requirements.txt
einops
lpips
tqdm
ptflops
pyYAML
thop
tifffile
torchstat
timm==0.5.4
numpy==1.22.0
imageio==2.13.5
matplotlib==3.5.1
opencv-python==4.5.5.62
Pillow==9.0.0
scipy==1.8.1
scikit-image==0.19.3
scikit-video==1.1.11
